requirements.txt
streamlit
yfinance
talib-binary